requirements.txt
django
djangorestframework
django_environ
django_cors_headers
